Add basic SEO metadata and sitemap
