+++ public/robots.txt
@@ -1,0 +1,3 @@
+User-agent: *
+Allow: /
+Sitemap: https://www.hilwantransport.ae/sitemap.xml
