+++ src/app/layout.tsx
@@ -1,27 +1,35 @@
 import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faWhatsapp } from "@fortawesome/free-brands-svg-icons";
 
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
 
 const whatsappNumber = "971588888302";
 const whatsappMessage = encodeURIComponent("Hello, I am interested in learning more about your crane rental, transport, and logistics services. Please provide more details.");
 const whatsappLink = `https://wa.me/${whatsappNumber}?text=${whatsappMessage}`;
 
 export const metadata: Metadata = {
+  metadataBase: new URL("https://www.hilwantransport.ae"),
   title: "Hilwan General Land Transport L.L.C.",
   description:
     "Hilwantransport.ae - Professional rental cranes, cargo transport by light and heavy trucks, and construction logistics in the UAE since 1994. Contact: +971 58 888 8302, hilwantransport@gmail.com.",
+  keywords: [
+    "Hilwan",
+    "crane rental",
+    "cargo transport",
+    "logistics",
+    "UAE",
+  ],
+  openGraph: {
+    title: "Hilwan General Land Transport L.L.C.",
+    description:
+      "Professional crane rental and heavy transport services across the UAE.",
+    url: "https://www.hilwantransport.ae",
+    siteName: "Hilwan General Land Transport L.L.C.",
+    images: ["/logo.jpeg"],
+    locale: "en_US",
+    type: "website",
+  },
 };
 
 export default function RootLayout({
@@ -31,7 +39,7 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body className={`${geistSans.variable} ${geistMono.variable}`}>
+      <body>
         {children}
         <a
           href={whatsappLink}
